+++ web/src/components/organisms/Settings/MyIntegrations/index.tsx
@@ -1,0 +1,13 @@
+import IntegrationCreationModal from "@reearth-cms/components/molecules/MyIntegrations/IntegrationCreationModal";
+import MyIntegrationsList from "@reearth-cms/components/molecules/MyIntegrations/MyIntegrationsList";
+
+const MyIntegrations: React.FC = () => {
+  return (
+    <>
+      <MyIntegrationsList />
+      <IntegrationCreationModal open={false} />
+    </>
+  );
+};
+
+export default MyIntegrations;
