+++ web/src/components/pages/MyIntegrations/index.tsx
@@ -1,15 +1,15 @@
 import { AuthenticationRequiredPage } from "@reearth-cms/auth";
 import Dashboard from "@reearth-cms/components/organisms/Dashboard";
-import MyIntegration from "@reearth-cms/components/organisms/Settings/MyIntegration";
+import MyIntegrations from "@reearth-cms/components/organisms/Settings/MyIntegrations";
 
-const MyIntegrationPage: React.FC = () => {
+const MyIntegrationsPage: React.FC = () => {
   return (
     <AuthenticationRequiredPage>
       <Dashboard defaultSelectedKeys={["my-integration"]}>
-        <MyIntegration />
+        <MyIntegrations />
       </Dashboard>
     </AuthenticationRequiredPage>
   );
 };
 
-export default MyIntegrationPage;
+export default MyIntegrationsPage;
